move inference into separate module
