+++ src/mozilla_sec_eia/ex_21/create_labeled_dataset.py
@@ -8,9 +8,9 @@ from pathlib import Path
 import pandas as pd
 
 from mozilla_sec_eia.utils.cloud import GCSArchive
+from mozilla_sec_eia.utils.layoutlm import normalize_bboxes
 from mozilla_sec_eia.utils.pdf import (
     get_pdf_data_from_path,
-    normalize_bboxes,
     pil_to_cv2,
     render_page,
 )
