+++ src/mozilla_sec_eia/utils/pdf.py
@@ -13,7 +13,7 @@ import numpy as np
 import pandas as pd
 from fitz import Rect
 from matplotlib import pyplot as plt
-from PIL import Image, ImageDraw, ImageFont
+from PIL import Image
 
 PDF_POINTS_PER_INCH = 72  # believe this is standard for all PDFs
 logger = logging.getLogger(__name__)
@@ -417,67 +417,3 @@ def _pil_img_from_pixmap(pix: fitz.Pixmap) -> Image.Image:
 
     img = Image.frombytes(mode, (pix.width, pix.height), pix.samples)
     return img
-
-
-def normalize_bboxes(txt_df, pg_meta_df):
-    """Normalize bboxes between 0 and 1000."""
-    txt_df["top_left_x_pdf"] = (
-        txt_df["top_left_x_pdf"] / pg_meta_df.width_pdf_coord.iloc[0] * 1000
-    )
-    txt_df["top_left_y_pdf"] = (
-        txt_df["top_left_y_pdf"] / pg_meta_df.height_pdf_coord.iloc[0] * 1000
-    )
-    txt_df["bottom_right_x_pdf"] = (
-        txt_df["bottom_right_x_pdf"] / pg_meta_df.width_pdf_coord.iloc[0] * 1000
-    )
-    txt_df["bottom_right_y_pdf"] = (
-        txt_df["bottom_right_y_pdf"] / pg_meta_df.height_pdf_coord.iloc[0] * 1000
-    )
-    return txt_df
-
-
-def unnormalize_box(bbox, width, height):
-    """Unnormalize bboxes for drawing onto an image."""
-    return [
-        width * (bbox[0] / 1000),
-        height * (bbox[1] / 1000),
-        width * (bbox[2] / 1000),
-        height * (bbox[3] / 1000),
-    ]
-
-
-def _iob_to_label(label):
-    label = label[2:]
-    if not label:
-        return "other"
-    return label
-
-
-def draw_boxes_on_img(
-    preds_or_labels,
-    boxes,
-    image,
-    width,
-    height,
-    font=ImageFont.load_default(),
-    unnormalize=False,
-):
-    """Draw bounding boxes on an image.
-
-    Useful for visualizing result of inference.
-    """
-    draw = ImageDraw.Draw(image)
-    label_color_lookup = {
-        "subsidiary": "green",
-        "loc": "red",
-        "own_per": "orange",
-    }
-    for pred_or_label, box in zip(preds_or_labels, boxes):
-        label = _iob_to_label(pred_or_label).lower()
-        if label == "other":
-            continue
-        if unnormalize:
-            box = unnormalize_box(box, width, height)
-        color = label_color_lookup[label]
-        draw.rectangle(box, outline=color)
-        draw.text((box[0] + 10, box[1] - 10), text=label, fill=color, font=font)
